% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A02.analyzeCorpus.R
\name{analyzeCorpus}
\alias{analyzeCorpus}
\title{analyzeCorpus}
\usage{
analyzeCorpus(korpus)
}
\arguments{
\item{korpus}{- the meta data for the corpora to be analyzed}
}
\value{
analysis - data frame of descriptive statistics
}
\description{
\code{analyzeCorpus} summarizes descriptive statistics fora corpus
}
\details{
This function takes as its parameters, the type of analysis (full or fast),
and the meta data for the corpus
}
\seealso{
Other text processing functions: \code{\link{cleanCorpus}},
  \code{\link{cleanDocument}}, \code{\link{extractLines}},
  \code{\link{getCorpus}}, \code{\link{getStats}},
  \code{\link{summarizeAnalysis}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
