% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U0A.analyzeLexicalFeatures.R
\name{analyzeLexicalFeatures}
\alias{analyzeLexicalFeatures}
\title{analyzeLexicalFeatures}
\usage{
analyzeLexicalFeatures(document, posTags)
}
\arguments{
\item{document}{- the document to tag in unlisted token format}

\item{posTags}{- data frame of pos tags and descriptions}
}
\value{
analysis - analysis of lexical density for the file
 \itemize{
  \item{avgVc}{Numeric average coefficient of variation across all POS tags}
  \item{chunkMatrix}{Data frame of POS tags for each document chunk}
  \item{featureMatrix}{Data frame of POS tag features for each document chunk}
  \item{featureStats}{Dataframe of descriptive statistics for each POS tag}
  \item{tagPairs}{Dataframe of token POS tag pairs}

 }
}
\description{
\code{analyzeLexicalFeatures} Analyzes lexical features for a document
}
\details{
This function takes as its parameters, the listed tokenized document
to be tagged and returns an analysis of lexical features for the document.
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
