% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C00.buildPilot.R
\name{buildPilot}
\alias{buildPilot}
\title{buildPilot}
\usage{
buildPilot(clean, pilot, design)
}
\arguments{
\item{clean}{List with meta data for clean corpus}

\item{pilot}{List with meta data for pilot corpus}

\item{design}{Dataframe containing the pilot corpus design}
}
\description{
\code{buildPilot} Builds pilot corpus for analysis
}
\details{
This function takes as its parameters, the meta data for the clean and
pilot corpora, and the pilot corpus design and  creates a pilot corpus
according to the design and stores it for downsream analysis.
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
