% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U00.checkDir.R
\name{checkDir}
\alias{checkDir}
\title{checkDir}
\usage{
checkDir(directory)
}
\arguments{
\item{directory}{- the directory to be checked}
}
\value{
number of files in the directory
}
\description{
\code{checkDir} creates a directory if it doesn't exist
}
\details{
This function takes a directory as a parameter, checks for its existence.
if the directory doesn't exist, it is created and the function returns a
zero value.  Otherwise, the number of files in the directory is returned.
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
