% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U07.chunkDocument.R
\name{chunkDocument}
\alias{chunkDocument}
\title{chunkDocument}
\usage{
chunkDocument(document, chunkSize)
}
\arguments{
\item{document}{- the document to be sampled}

\item{chunkSize}{- the number of sentences per chunk}
}
\value{
chunks - list of chunks of equal number of sentences
}
\description{
This function takes as its parameters, a document in unlisted sentence or
word tokenized format and the chunk size in terms of numbers of tokens,
and returns the document in a list of chunks of equal size.
}
\author{
John James
}
