% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01.cleanData.R
\name{cleanCorpus}
\alias{cleanCorpus}
\title{cleanCorpus}
\usage{
cleanCorpus(raw, clean)
}
\arguments{
\item{raw}{- the meta data for the reshaped raw corpus}

\item{clean}{- the meta data for the clean corpus}
}
\description{
\code{cleanCorpus} cleans the HC Corpus
}
\details{
This function cleans the raw corpus data and stores it in the designated
directory
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
