% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01.cleanCorpus.R
\name{cleanDocument}
\alias{cleanDocument}
\title{cleanDocument}
\usage{
cleanDocument(rawDocument)
}
\arguments{
\item{rawDocument}{- the meta data and content for the file to be analyzed}
}
\value{
cleanDocument Cleaned text document in unlisted vector format.
}
\description{
\code{cleanDocument} cleans the HC Corpus document
}
\details{
This function reads a corpus document and performs a series of conditioning,
reshaping, normalization, and clean up tasks.
The conditioning tasks include:
\itemize{
 \item{Control}{Replace UTF-8 control characters with spaces}
 \item{Quotations}{Convert non-standard quotations to single quote}
 \item{Encoding}{Convert document from UTF-8 to ASCII enconding}
 }

The reshaping task uses the quanteda package
 \url{https://cran.r-project.org/web/packages/quanteda/quanteda.pdf} to
 reshape the corpus documents into sentences

The normalization tasks include the following:
 \itemize{
  \item{control}{Remove ASCII control characters}
  \item{nonprint}{Remove ASCII non-printable characters}
  \item{emails}{Remove email addresses}
  \item{urls}{Remove urls}
  \item{hashtags}{Remove twitter hashtags}
  \item{control}{Remove ASCII control characters}
  \item{digits}{Remove digits}
  \item{punct}{Remove punctuation except the apostrophe}
  \item{longWords}{Remove words 40 characters or longer}
  \item{profanity}{Remove profanity}
  \item{correct}{Correct contractions and common misspellings}#'
 }

Finally, the clean up tasks include:
 \itemize{
  \item{whiteSpace}{Remove extra whitespace from documents}
  \item{punct}{Remove stray apostrophes and punctuation}
  \item{emptySent}{Remove empty sentences}
 }
}
\seealso{
Other text processing functions: \code{\link{analyzeCorpus}},
  \code{\link{cleanCorpus}}, \code{\link{extractLines}},
  \code{\link{getCorpus}}, \code{\link{getStats}},
  \code{\link{summarizeAnalysis}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
