% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B00.createVGC.R
\name{createGrowthObject}
\alias{createGrowthObject}
\title{createGrowthObject}
\usage{
createGrowthObject(document, type = "fast", samples = 1000)
}
\arguments{
\item{document}{The meta data for the document being analyzed}

\item{type}{- c('fast', 'full')}

\item{samples}{- the number of rows in the growth document}
}
\value{
vgc - vocabulary growth curve data
}
\description{
\code{createGrowthObject} Creates vocabulary growth object for document
}
\details{
This function takes as its parameter, the meta data for the document to be
analyzed, a string indicating the type of analysis to perform c('full', 'fast),
and the number of samples and returns  and returns vocabulary growth data
at increments designated by "samples". For full analysis, the vgc function
from the languageR package is used
}
\seealso{
\code{\link{createVGC}}
 \code{\link{createZipf}}
 \code{\link{createZipfObject}}
 \code{\link[quanteda]{tokenize}}
 \code{\link[languageR]{growth.fnc}}

Other vocabulary growth functions: \code{\link{createVGC}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
