% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B00.createVGC.R
\name{createVGC}
\alias{createVGC}
\title{createVGC}
\usage{
createVGC(korpus, type = "fast")
}
\arguments{
\item{korpus}{- the meta data for the korpus being analyzed}

\item{type}{= c('fast', 'full)}
}
\value{
growth - list of growth objects
}
\description{
\code{createGrowthObject} Creates vocabulary growth data for a corpus
}
\details{
This function takes as its parameter the korpus meta data and type of
analysis to conduct and returns vocabulary growth data indicating how the
vocabulary grows in terms of new unique tokens as the V
}
\seealso{
Other vocabulary growth functions: \code{\link{createGrowthObject}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
