% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B01.createZipf.R
\name{createZipf}
\alias{createZipf}
\title{createZipf}
\usage{
createZipf(korpus, vgc)
}
\arguments{
\item{korpus}{- meta data for the corpus being analyzed}

\item{vgc}{- the vocabulary growth curve data}
}
\value{
zipfObjects - the zipf object used for analysis and plotting
}
\description{
\code{createZipf} Prepares zipf objects for a corpus
}
\details{
This function takes as its parameters, the corpus meta data and a
vocabulary growth curve and returns a zip object used for downstream
analysis. The zipf object includes:
\itemize{
 \item{nTokens}{Number of tokens in the document}
 \item{docTfl}{Document term frequency list}
 \item{docSpc}{Document frequency spectrum object}
 \item{docLnre}{Large number of rare elements (LNRE) object}
 \item{docEmpVgc}{Empirical (observed) vocabulary growth curve}
 \item{docIntVgc}{Interpolated (smoothed) vocabulary growth curve}
 \item{goodness}{LNRE goodness of fit}#'
}
}
\seealso{
\code{\link{createVGC}}
         \code{\link{createGrowthObject}}
         \code{\link{createZipfObject}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
