% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B01.createZipf.R
\name{createZipfObject}
\alias{createZipfObject}
\title{createZipfObject}
\usage{
createZipfObject(document, vgc, numSamples = 1000)
}
\arguments{
\item{document}{- the document and meta data to be analyzed}

\item{vgc}{- the vocabulary growth curve data}

\item{numSamples}{- the number of samples, default = 1000}
}
\value{
zipfObjects - the zipf object used for analysis and plotting
}
\description{
\code{createZipfObject} Prepares zipf objects for a document
}
\details{
This function takes as its parameters, the list containing the document
and its meta data, a vocabulary growth curve, the number of samples
(default = 1000),  and returns a zip object used for downstream analysis.
The zipf object includes:
\itemize{
 \item{nTokens}{Number of tokens in the document}
 \item{docTfl}{Document term frequency list}
 \item{docSpc}{Document frequency spectrum object}
 \item{docLnre}{Large number of rare elements (LNRE) object}
 \item{docEmpVgc}{Empirical (observed) vocabulary growth curve}
 \item{docIntVgc}{Interpolated (smoothed) vocabulary growth curve}
 \item{goodness}{LNRE goodness of fit}#'
}
}
\seealso{
\code{\link{createVGC}}
         \code{\link{createGrowthObject}}
         \code{\link{createZipf}}
         \code{\link[languageR]{text2spc.fnc}}
         \code{\link[zipfR]{lnre}}
         \code{\link[zipfR]{lnre.spc}}
         \code{\link[zipfR]{N}}
         \code{\link[zipfR]{V}}
         \code{\link[zipfR]{EV}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
