% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B06.designModelCorpora.R
\name{designModelCorpora}
\alias{designModelCorpora}
\title{designModelCorpora}
\usage{
designModelCorpora(sampleSizeEstimate, registerSizeEstimate, samplingUnit,
  analysis)
}
\arguments{
\item{sampleSizeEstimate}{- the vocabulary-based estimate}

\item{registerSizeEstimate}{- the lexical feature based estimate}

\item{samplingUnit}{- the sampling unit}

\item{analysis}{- the clean data analysis}
}
\value{
alpha - the corpus design for alpha training set.
}
\description{
\code{designModelCorpora} Designs the model corpora
}
\details{
This function takes as its parameters, the the vocabulary based sample size
estimate from \code{\link{estimateSampleSize}}, the register size estimates
from \code{\link{estimateRegisterSize}}, the sampling unit estimate
from \code{\link{estimateSamplingUnit}}, and the clean data analysis from
\code{\link{analyzeData}}, and produces the model corpora design including
sizes for four training sets, a validation set and a test set.
}
\author{
John James
}
