% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B04.estimateCorpusSize.R
\name{estimateCorpusSize}
\alias{estimateCorpusSize}
\title{estimateCorpusSize}
\usage{
estimateCorpusSize(korpus, sampleSize = 2000, numSamples = 100)
}
\arguments{
\item{korpus}{List containing the corpus meta data}

\item{sampleSize}{Numeric indicating the sampling unit size}

\item{numSamples}{Numeric indicating the number of samples to analyize}
}
\value{
corpusSize List cointaining:
\itemize{
 \item{n}{Corpus sample size estimate}
 \item{posAnalysis}{POS tag distribution analysis}
}
}
\description{
\code{estimateCorpusSize} Estimates corpus size based upon lexical features
}
\details{
This function takes as its parameters, the korpus meta data and the POS tags
selected for this analysis, the returns an estimate of total corpus size
based upon the distribution of lexical  features per n000-word samples of
the text. This analysis is based upon Representativeness in Corpus Design
Biber 1993
\url{https://www.researchgate.net/publication/31460364_Representativeness_in_Corpus_Design}
}
\seealso{
\code{\link{analyzeLexicalFeatures}}

Other sample size estimate functions: \code{\link{estimateRegisterSize}},
  \code{\link{estimateSampleSize}},
  \code{\link{estimateSamplingUnit}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
