% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B05.estimateRegisterSize.R
\name{estimateRegisterSize}
\alias{estimateRegisterSize}
\title{estimateRegisterSize}
\usage{
estimateRegisterSize(korpus, corpusSize, samplingUnit, sampleSize = 2000,
  numSamples = 100)
}
\arguments{
\item{korpus}{List containing the meta data for the corpus}

\item{corpusSize}{List containing teh corpus size estimate from \code{\link{estimateCorpusSize}}}

\item{samplingUnit}{- List containing the sampling unit estimate from \code{\link{estimateSamplingUnit}}}

\item{sampleSize}{Integer indicating sampling unit size}

\item{numSamples}{Integer indicating the number of samples to analyze}
}
\value{
registerSizes Dataframe containing
\itemize{
\item{Register}{String indicating the register name}
\item{Base}{Numeric indicating base allocation of samples allocated to all registers}
\item{Avg Vc}{Numberic average coefficient of variation across all POS tags}
\item{lambda}{Numeric factor multiplied by Avg Vc to calculate proportional allocation}
\item{Proportion}{Numeric indicating the proportional allocation of samples to registers}
\item{Num Samples}{Integer indicating Base + Proportion for each register}
\item{Sample Length}{Integer indicating length of each sample in tokens}
\item{Sample Size}{Integer = Num Samples * Sample Length}
}
}
\description{
\code{estimateRegisterSize} Estimates the size of individual registers
}
\details{
This function takes as its parameters, the meta data for the korpus,
the corpus sample size estimate from \code{\link{estimateCorpusSize}},
the sampling unit estimate from \code{\link{estimateSamplingUnit}}, the
POS tags, sample size, and the number of samples and returns an estimate of
register size for each register based upon the distribution of lexical
features per 2000-word samples of the text. This analysis is based upon
Biber's 1993 Representativeness in Corpus Design.
}
\seealso{
Other sample size estimate functions: \code{\link{estimateCorpusSize}},
  \code{\link{estimateSampleSize}},
  \code{\link{estimateSamplingUnit}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
