% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/B02.estimateSampleSize.R
\name{estimateSampleSize}
\alias{estimateSampleSize}
\title{estimateSampleSize}
\usage{
estimateSampleSize(korpus, analysis)
}
\arguments{
\item{korpus}{- the meta data for the HC Corpus}

\item{analysis}{- the analysis of the HC Corpus}
}
\value{
estimate - list of data frames containing coverage estimates
}
\description{
\code{estimateSampleSize} Estimates sample sizes for various OOV rates
}
\details{
This function takes as its parameters, the corpus meta data and the analysis
of the clean corpus then uses LNRE models from the zipfR package to
calculate and evaluate various samples sizes and associated OOV Rates.
}
\seealso{
\code{\link[languageR]{text2spc.fnc}}
  \code{\link[zipfR]{lnre}}
  \code{\link[zipfR]{lnre.spc}}
  \code{\link[zipfR]{N}}
  \code{\link[zipfR]{V}}
  \code{\link[zipfR]{EV}}

Other sample size estimate functions: \code{\link{estimateCorpusSize}},
  \code{\link{estimateRegisterSize}},
  \code{\link{estimateSamplingUnit}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
