% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U0B.extractLines.R
\name{extractLines}
\alias{extractLines}
\title{extractLines}
\usage{
extractLines(document, strings)
}
\arguments{
\item{document}{List of texts to be cleaned}

\item{strings}{Vector of strings to be removed}
}
\value{
cleanDocument List of text without lines containing profanity
}
\description{
\code{extractLines} removes text lines which include designated strings
}
\details{
This function takes as its parameters, a list or vector of texts and a
vector of strings to be excluded from the text, then removes the lines
containing the string(s) and returns the cleaned document
to the calling environment.
}
\seealso{
Other text processing functions: \code{\link{analyzeCorpus}},
  \code{\link{cleanCorpus}}, \code{\link{cleanDocument}},
  \code{\link{getCorpus}}, \code{\link{getStats}},
  \code{\link{summarizeAnalysis}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
