% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A00.getCorpus.R
\name{getCorpus}
\alias{getCorpus}
\title{getCorpus}
\usage{
getCorpus(force = FALSE)
}
\arguments{
\item{force}{= Logical. Force download and overwrite local copy if TRUE}
}
\value{
sysData Internal data and data structures
}
\description{
\code{getCorpus} obtains the HC Corpora from the website
}
\details{
The function obtains the HC Corpora data from its online source if the
 document doesn't already exist
}
\seealso{
Other text processing functions: \code{\link{analyzeCorpus}},
  \code{\link{cleanCorpus}}, \code{\link{cleanDocument}},
  \code{\link{extractLines}}, \code{\link{getStats}},
  \code{\link{summarizeAnalysis}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
