% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A00.getData.R
\name{getData}
\alias{getData}
\title{getData}
\usage{
getData(force = FALSE)
}
\arguments{
\item{force}{= Logical. Force download and overwrite local copy if TRUE}
}
\value{
sysData Internal data and data structures
}
\description{
\code{getData} obtains the HC Corpora from the website
}
\details{
The function obtains the HC Corpora data from its online source if the
 document doesn't already exist
}
\seealso{
Other data processing functions: \code{\link{cleanCorpus}},
  \code{\link{cleanDocument}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
