% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A02.analyzeData.R
\name{getFeatures}
\alias{getFeatures}
\title{getFeatures}
\usage{
getFeatures(type, document)
}
\arguments{
\item{type}{Character string c('full', 'fast')}

\item{document}{- the document content to be analyzed}
}
\value{
analysis - a list containing the following:
  \itemize{
   \item{featureVector}{List of descriptive statistics}
   \item{featureSamples}{List of samples of select lexical features}
   \item{featureTables}{List of data frames of features and counts}
}
}
\description{
\code{getFeatures} returns descriptive statistics and lexical features
}
\details{
This function takes as its parameters, the type of analysis and the
document content to be analyzed and returns descriptive statistics and
lexical features for and in the document.
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
