% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U01.loadObject.R
\name{loadObject}
\alias{loadObject}
\title{loadObject}
\usage{
loadObject(object)
}
\arguments{
\item{object}{A list containing a directory and filename}
}
\value{
object R object loaded
}
\description{
\code{loadObject} loads an R object from an RDA file into memory
}
\details{
This function takes as its parameter a list containing a directory and
filename. It loads the associated object into memory and returns it to the
calling environment.
}
\seealso{
Other io functions: \code{\link{readFile}},
  \code{\link{saveFile}}, \code{\link{saveObject}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
