% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C02.nGramCorpus.R
\name{nGramCorpus}
\alias{nGramCorpus}
\title{nGramCorpus}
\usage{
nGramCorpus(korpus)
}
\arguments{
\item{korpus}{- the meta data for pilot corpus}
}
\description{
\code{nGramCorpus} Creates unigrams, bigrams, trigrams, and quadgrams
}
\details{
This function takes a corpus meta data as an argument and  prepares unigrams,
bigrams, trigrams, and quadgrams for a corpus and stores it in the directory
designated in the corpus meta data.
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
