% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U05.parallelizeTask.R
\name{parallelizeTask}
\alias{parallelizeTask}
\title{parallelizeTask}
\usage{
parallelizeTask(task, ...)
}
\arguments{
\item{task}{- the function to parallelize}

\item{...}{- other parameters required by task}
}
\value{
r - return from task
}
\description{
\code{parallelizeTask} parallelize task
}
\details{
This generic function used to parallelize a task (when possible)
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
