% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Platform.R
\docType{class}
\name{Platform}
\alias{Platform}
\title{Platform}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Platform
}
\value{
Object of \code{\link{R6Class}} with methods for managing environments
}
\description{
\code{Platform} Class the creates and manages environments.
}
\details{
This class creates and manages environments.  An environment is essentially
a directory in which project data resides.  Multiple environments can
be created to provide multiple data environments. This is especially
useful for projects with large data sets or computationally expense
analyses.
}
\section{Methods}{

\describe{
 \item{\code{new(pName, pDesc)}}{Initializes the platform}
 \item{\code{createEnv(envName, envDesc, envPath, envStatus)}}{Creates an environment}
 \item{\code{archiveEnv(environment)}}{Archives an environment}
 \item{\code{listEnv()}}{Provides a data frame of environments in the platform}
}
}

\section{Active Bindings}{

\describe{
 \item{\code{currentEnv}}{Gets and sets the current environment}
}
}

\examples{
\dontrun{
predictor$new(pName = "predictor", pDesc = "Platform for predictor project")
predictor$createEnv(envName = "test", envDesc = "Test Environment",
                    envPath = "test", envStatus = TRUE)
predictor$archiveEnv(name = 'test')
predictor$listEnv()
predictor$currentEnv
predictor$currentEnv <- "production"
}

}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
