% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U03.readFile.R
\name{readFile}
\alias{readFile}
\title{readFile}
\usage{
readFile(object)
}
\arguments{
\item{object}{List containing directory and filename}
}
\value{
fileData The csv, dic, or text data
}
\description{
\code{readFile} reads a csv, dic, or text file into memory
}
\details{
The function takes a directory and file name as s parameter, reads the
file and returns the object to the calling environment.
}
\seealso{
Other io functions: \code{\link{loadObject}},
  \code{\link{saveFile}}, \code{\link{saveObject}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
