% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U06.sampleData.R
\name{sampleData}
\alias{sampleData}
\title{sampleData}
\usage{
sampleData(document, numChunks, chunkSize, format = "lv")
}
\arguments{
\item{document}{Vector containing a text document}

\item{numChunks}{Integer indicating number of chunks or samples to return}

\item{chunkSize}{Integer indicating the size chunk/sample size in percent or
number of tokens}

\item{format}{- format of the data to be returned values are:
lv = list of character vectors
ls = list of character strings
v = unlisted character vector
s = unlist character string}
}
\value{
samples - the samples in a contiguous vector of tokens
}
\description{
\code{sampleData} Returns samples from a corpus document
}
\details{
This function takes as its parameters, a document in unlisted tokenized
format, the number of chunks, the chunk size, and the format of the returned
document, then returns a sample document according to parameters.
}
\author{
John James
}
