% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U04.saveFile.R
\name{saveFile}
\alias{saveFile}
\title{saveFile}
\usage{
saveFile(object)
}
\arguments{
\item{object}{- a list containing:
\itemize{
   \item{directory}{Directory into which the data is to be stored}
   \item{fileName}{File name given to the file to be saved}
   \item{data}{File contents}#'
   }}
}
\description{
\code{saveFile} saves a csv, dic, or text file to disk
}
\details{
This function function takes as a parameter, file and meta data and
saves the file in accordance with the file meta data (directory, filename).
}
\seealso{
Other io functions: \code{\link{loadObject}},
  \code{\link{readFile}}, \code{\link{saveObject}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
