% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U02.saveObject.R
\name{saveObject}
\alias{saveObject}
\title{saveObject}
\usage{
saveObject(object)
}
\arguments{
\item{object}{List containing:
\item{directory}{Directory into which the data is to be stored}
\item{objName}{Object name given to the R object created}
\item{fileName}{File name for the R object created}}
}
\value{
object - R object saved
}
\description{
\code{saveObject} saves an R object in rda format.
}
\details{
This function function takes as a parameter, a list containing both data and
meta data (directory, object name, filename), and saves the data in an
RDA file according to the metaData.
}
\seealso{
Other io functions: \code{\link{loadObject}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
