% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U08.tagChunk.R
\name{tagChunk}
\alias{tagChunk}
\title{tagChunk}
\usage{
tagChunk(chunk, chunkNum, sentAnnotator, wordAnnotator, posAnnotator, posTags)
}
\arguments{
\item{chunk}{- a chunk (list) of text to be tagged}

\item{chunkNum}{- the chunk number}

\item{sentAnnotator}{- the openNLP sentence annotator}

\item{wordAnnotator}{- the openNLP word annotator}

\item{posAnnotator}{- the openNLP POS annotator}

\item{posTags}{- pos tags and descriptions}
}
\value{
tags - 1 row data frame with frequencies for each feature
}
\description{
\code{tagChunk} POS tags chunk of text and returns POS counts for the chunk
}
\details{
This function takes as its parameter, the a list of text to be tagged,
the chunk number, the POS tags and the sentence, word, and POS annotators
from the openNLP package and returns a one-row data frame containing
frequences of POS tags for the chunk.
}
\seealso{
Other POS Tagging Functions: \code{\link{tagDocument}},
  \code{\link{tagSentence}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
