% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C01.tagCorpus.R
\name{tagCorpus}
\alias{tagCorpus}
\title{tagCorpus}
\usage{
tagCorpus(korpus)
}
\arguments{
\item{korpus}{- meta data for the corpus to be tagged}
}
\description{
\code{buildPilot} POS tag a corpus
}
\details{
This function takes as its parameter, the meta data for a corpus and performs
POS tagging of the corpus documents and stores the results in the
document /  POS directory
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
