% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U09.tagDocument.R
\name{tagDocument}
\alias{tagDocument}
\title{tagDocument}
\usage{
tagDocument(document)
}
\arguments{
\item{document}{- the document object with meta data and content}
}
\value{
posData -  a list containing the tagged text and the Word/Pos pairs
}
\description{
\code{tagDocument} POS tags a document and returns tags and word/tag pairs
}
\details{
This function takes as its parameter, a corpus document including the meta
data and the content in the form of vectors of sentences, and returns the
tagged sentences as well as WORD/POS pairs.
}
\seealso{
Other POS Tagging Functions: \code{\link{tagChunk}},
  \code{\link{tagSentence}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
