% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/U09.tagDocument.R
\name{tagSentence}
\alias{tagSentence}
\title{tagSentence}
\usage{
tagSentence(sentence, sentAnnotator, wordAnnotator, posAnnotator)
}
\arguments{
\item{sentence}{Vector of text comprising a sentence}

\item{sentAnnotator}{- the openNLP sentence annotator}

\item{wordAnnotator}{- the openNLP word annotator}

\item{posAnnotator}{- the openNLP POS annotator}
}
\description{
\code{tagSentence} POS tags a sentence and returns tags and word/tag pairs
}
\details{
This function takes as its parameters, a sentence to be tagged, as well as
the sentence, word and pos tag annotators from the openNLP package and returns
the POS tags and word/tag pairs.
}
\seealso{
Other POS Tagging Functions: \code{\link{tagChunk}},
  \code{\link{tagDocument}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
