% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C04.verifyLinguistics.R
\name{verifyCorpusLexicalFeatures}
\alias{verifyCorpusLexicalFeatures}
\title{verifyCorpusLexicalFeatures}
\usage{
verifyCorpusLexicalFeatures(korpus, chunks, chunkSize)
}
\arguments{
\item{korpus}{List containing meta data for the corpus}

\item{chunks}{Integer indicating number of chunks to sample}

\item{chunkSize}{Integer indicating sizes of chunk in sentences}
}
\value{
analysis - the lexical feature distribution analysis.
}
\description{
\code{getCorpus} Verifies lexical features distribution of a corpus
}
\details{
This function takes as its parameters, a list containing the korpus
meta data, and two integers, one indicating the number of chunks to sample
the other indicating the size of each chunk in sentences. The function then
reads, samples, and analyzes the lexical feature distribution for the
corpus.
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
