% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C04.verifyLinguistics.R
\name{verifyLinguistics}
\alias{verifyLinguistics}
\title{verifyLinguistics}
\usage{
verifyLinguistics(fullCorpus, sampleCorpus, chunks = 100, chunkSize = 2000)
}
\arguments{
\item{fullCorpus}{List containing the meta data for the HC Corpus}

\item{sampleCorpus}{List containing the meta data for the sampleCorpus corpus}

\item{chunks}{The number of chunks to sample}

\item{chunkSize}{The number of sentences in each chunk sampled}
}
\value{
features A list containing:
\itemize{
 \item{analyses}{Lexical feature analysis for both corpora}
 \item{means}{The mean numbers of features per sample}
 \item{x2}{Chi-squared p-value}
}
}
\description{
\code{getCorpus} Compares distribution of lexical features between 2 corpora
}
\details{
This function takes as its parameters, the meta data for the full corpus
and the sample corpus and returns comparative statistics of the
distribution of lexical features.
}
\author{
John James
}
