% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/C03.verifyVocabulary.R
\name{verifyVocabulary}
\alias{verifyVocabulary}
\title{verifyVocabulary}
\usage{
verifyVocabulary(fullCorpus, sampleCorpus)
}
\arguments{
\item{fullCorpus}{- the meta data for the full corpus}

\item{sampleCorpus}{- the meta data for the sample corpus}
}
\value{
oovRates - the OOV Rate for each corpus
}
\description{
\code{verifyVocabulary} Verifies vocabulary coverage for a corpus
}
\details{
This function takes as its parameter, the meta data for a full corpus and
for the sample corpus and checks the out of vocabulary (OOV) rates of
a sample corpus vis-a-vis a "full corpus".  Words in the full corpus not
found in the sample corpus are "OOV"
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
