% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A02.analyzeData.R
\name{analyzeData}
\alias{analyzeData}
\title{analyzeData}
\usage{
analyzeData(korpus)
}
\arguments{
\item{korpus}{- the meta data for the corpora to be analyzed}
}
\value{
analysis - data frame of descriptive statistics
}
\description{
\code{analyzeData} summarizes descriptive statistics fora corpus
}
\details{
This function takes as its parameters, the type of analysis (full or fast),
and the meta data for the corpus
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
