% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01.cleanData.R
\name{cleanCorpus}
\alias{cleanCorpus}
\title{cleanCorpus}
\usage{
cleanCorpus(raw, clean)
}
\arguments{
\item{raw}{- the meta data for the reshaped raw corpus}

\item{clean}{- the meta data for the clean corpus}
}
\description{
\code{cleanCorpus} cleans the HC Corpus
}
\details{
This function takes as its parameters, the meta data for the raw and clean
corpora, then performs normalization and cleaning tasks and stores it in
the directory designated in the meta data.
}
\seealso{
Other data processing functions: \code{\link{cleanDocument}},
  \code{\link{getData}}
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
