% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A01.cleanData.R
\name{cleanDocument}
\alias{cleanDocument}
\title{cleanDocument}
\usage{
cleanDocument(rawDocument)
}
\arguments{
\item{rawDocument}{- the meta data and content for the file to be analyzed}
}
\value{
cleanDocument Cleaned text document in unlisted vector format.
}
\description{
\code{cleanDocument} cleans the HC Corpus document
}
\details{
This function takes as its parameters, the meta data for the raw corpus
document and performs a series of normalization and data cleaning tasks
and returns the cleaned document to the calling environment.
}
\author{
John James
}
