% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00.Platform.R
\docType{class}
\name{Platform}
\alias{Platform}
\title{Platform}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Platform
}
\value{
Object of \code{\link{R6Class}} with methods for managing environments
}
\description{
\code{Platform} Class the creates and manages environments.
}
\details{
This class creates and manages environments.  An environment is essentially
a directory in which project data resides.  Multiple environments can
be created to provide multiple versions of the data. This is especially
useful for projects with large data sets. Users can create environments
to contain smaller versions of the data sets for testing purposes.
}
\section{Methods}{

\describe{
 \item{\code{new(name, desc)}}{Initializes the platform}
 \item{\code{printPlatform()}}{Prints the platform object definition}
 \item{\code{createEnv(name, desc, path, current)}}{Creates an environment}
 \item{\code{archiveEnv(name)}}{Archives an environment}
 \item{\code{existsEnv(name)}}{Checks existence of an environment}
 \item{\code{listEnv()}}{Provides a data frame of environments in the platform}
}
}

\section{Active Bindings}{

\describe{
 \item{\code{currentEnv}}{Gets and sets the current environment}
}
}

\examples{
\dontrun{
Platform$new(name = "predictor", desc = "Predictor object")
predictor$printPlatform()
predictor$createEnv(name = "test", desc = "Test Environment", path = "test", current = TRUE)
predictor$archiveEnv(name = 'test')
predictor$existsEnv(name = 'production')
predictor$listEnv()
predictor$currentEnv
predictor$currentEnv <- "production"
}

}
\author{
John James, \email{j2sdatalab@gmail.com}
}
\keyword{datasets}
