% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A02.analyzeData.R
\name{summarizeAnalysis}
\alias{summarizeAnalysis}
\title{summarizeAnalysis}
\usage{
summarizeAnalysis(type, korpus, featureMatrix, featureSamples, regex)
}
\arguments{
\item{type}{Character string indicating type of analysis c("full", "fast:)}

\item{korpus}{- the korpus and its meta data}

\item{featureMatrix}{- the feature matrix for the 3 registers}
}
\value{
totals - a data frame containing total feature counts for the corpus
}
\description{
\code{summarizeAnalysis} summarizes descriptive statistics and lexical features
}
\details{
This function takes as its parameter, the type of analysis to be conducted,
the korpus meta data, and the feature matrix and summarizes the descriptive
statistics and lexical features at the corpus level.
}
\author{
John James
}
