% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A02.analyzeData.R
\name{summarizeAnalysis}
\alias{summarizeAnalysis}
\title{summarizeAnalysis}
\usage{
summarizeAnalysis(korpus, stats)
}
\arguments{
\item{korpus}{The corpus documents}

\item{stats}{The descriptive stats for the three registers.}
}
\value{
totals A data frame of descriptive statistics and totals
}
\description{
\code{summarizeAnalysis} summarizes descriptive statistics
}
\details{
This function takes as its parameters, the corpus documents and the
descriptive stats for the registers and produces summary statistics
at the corpus level.
}
\author{
John James, \email{j2sdatalab@gmail.com}
}
